Created bars-package
