+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - Extracted NumFactory as source of numbers with defined precision
 - Replaced `ZonedDateTime` with `Instant`
 - Renamed `FixedDecimalIndicator` with `FixedNumIndicator`
+- Moved `BaseBar`, `BaseBarBuilder` and `BaseBarBuilderFactory` to `bars`-package
 
 ### Fixed
 - Fixed `BaseBar.toString()` to avoid `NullPointerException` if any of its property is null
@@ -30,6 +31,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - added `KRIIndicator`
 - Added constructor with `amount` for  `EnterAndHoldCriterion`
 - Added constructor with `amount` for  `VersusEnterAndHoldCriterion`
+- Added `TickBarBuilder` to `bars`-package
+- Added `VolumeBarBuilder` to `bars`-package
 
 ## 0.17 (released September 9, 2024)
