+++ ta4j-core/src/main/java/org/ta4j/core/bars/VolumeBarBuilderFactory.java
@@ -21,12 +21,27 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core;
+package org.ta4j.core.bars;
 
-class BaseBarBuilderFactory implements BarBuilderFactory {
+import org.ta4j.core.BarBuilder;
+import org.ta4j.core.BarBuilderFactory;
+import org.ta4j.core.BarSeries;
+
+public class VolumeBarBuilderFactory implements BarBuilderFactory {
+
+    private final int volumeThreshold;
+    private VolumeBarBuilder barBuilder;
+
+    public VolumeBarBuilderFactory(final int volumeThreshold) {
+        this.volumeThreshold = volumeThreshold;
+    }
 
     @Override
-    public BaseBarBuilder createBarBuilder(BarSeries series) {
-        return new BaseBarBuilder(series.numFactory()).bindTo(series);
+    public BarBuilder createBarBuilder(final BarSeries series) {
+        if (this.barBuilder == null) {
+            this.barBuilder = new VolumeBarBuilder(series.numFactory(), this.volumeThreshold).bindTo(series);
+        }
+
+        return this.barBuilder;
     }
 }
