+++ ta4j-core/src/test/java/org/ta4j/core/BarSeriesTest.java
@@ -41,6 +41,7 @@ import java.util.stream.IntStream;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.ta4j.core.bars.BaseBarBuilder;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.indicators.helpers.HighPriceIndicator;
