+++ ta4j-core/src/test/java/org/ta4j/core/BarTest.java
@@ -36,6 +36,7 @@ import java.time.ZonedDateTime;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.ta4j.core.bars.BaseBarBuilder;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
