+++ ta4j-core/src/test/java/org/ta4j/core/bars/BaseBarBuilderTest.java
@@ -21,57 +21,52 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core;
+package org.ta4j.core.bars;
 
 import static org.junit.Assert.assertEquals;
-import static org.ta4j.core.TestUtils.assertNumEquals;
 
-import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.Instant;
+
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import org.ta4j.core.BarSeries;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-@RunWith(Parameterized.class)
-public class BaseBarConvertibleBuilderTest extends AbstractIndicatorTest<BarSeries, Num> {
+public class BaseBarBuilderTest extends AbstractIndicatorTest<BarSeries, Num> {
 
-    public BaseBarConvertibleBuilderTest(NumFactory numFactory) {
+    public BaseBarBuilderTest(NumFactory numFactory) {
         super(numFactory);
     }
 
     @Test
-    public void testBuildBigDecimal() {
+    public void testBuildBar() {
 
         final Instant beginTime = Instant.parse("2014-06-25T00:00:00Z");
         final Instant endTime = Instant.parse("2014-06-25T01:00:00Z");
         final Duration duration = Duration.between(beginTime, endTime);
 
-        final var series = new BaseBarSeriesBuilder().withNumFactory(numFactory).build();
-        final var bar = series.barBuilder()
-                .timePeriod(duration)
+        final BaseBar bar = new BaseBarBuilder(numFactory).timePeriod(duration)
                 .endTime(endTime)
-                .openPrice(BigDecimal.valueOf(101.0))
-                .highPrice(BigDecimal.valueOf(103))
-                .lowPrice(BigDecimal.valueOf(100))
-                .closePrice(BigDecimal.valueOf(102))
+                .openPrice(numOf(101))
+                .highPrice(numOf(103))
+                .lowPrice(numOf(100))
+                .closePrice(numOf(102))
                 .trades(4)
-                .volume(BigDecimal.valueOf(40))
-                .amount(BigDecimal.valueOf(4020))
+                .volume(numOf(40))
+                .amount(numOf(4020))
                 .build();
 
         assertEquals(duration, bar.getTimePeriod());
         assertEquals(beginTime, bar.getBeginTime());
         assertEquals(endTime, bar.getEndTime());
-        assertNumEquals(numOf(101.0), bar.getOpenPrice());
-        assertNumEquals(numOf(103), bar.getHighPrice());
-        assertNumEquals(numOf(100), bar.getLowPrice());
-        assertNumEquals(numOf(102), bar.getClosePrice());
+        assertEquals(numOf(101), bar.getOpenPrice());
+        assertEquals(numOf(103), bar.getHighPrice());
+        assertEquals(numOf(100), bar.getLowPrice());
+        assertEquals(numOf(102), bar.getClosePrice());
         assertEquals(4, bar.getTrades());
-        assertNumEquals(numOf(40), bar.getVolume());
-        assertNumEquals(numOf(4020), bar.getAmount());
+        assertEquals(numOf(40), bar.getVolume());
+        assertEquals(numOf(4020), bar.getAmount());
     }
 }
