+++ ta4j-core/src/test/java/org/ta4j/core/bars/VolumeBarBuilderTest.java
@@ -21,7 +21,8 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core;
+
+package org.ta4j.core.bars;
 
 import static org.junit.Assert.assertEquals;
 import static org.ta4j.core.TestUtils.assertNumEquals;
@@ -30,12 +31,13 @@ import java.time.Duration;
 import java.time.Instant;
 
 import org.junit.Test;
+import org.ta4j.core.BaseBarSeriesBuilder;
 
-public class TickBarBuilderTest {
+public class VolumeBarBuilderTest {
 
     @Test
     public void add() {
-        final var series = new BaseBarSeriesBuilder().withBarBuilderFactory(new TickBarBuilderFactory(5)).build();
+        final var series = new BaseBarSeriesBuilder().withBarBuilderFactory(new VolumeBarBuilderFactory(4)).build();
         final var now = Instant.now();
         series.barBuilder().timePeriod(Duration.ofDays(1)).endTime(now).closePrice(1).volume(1).add();
         series.barBuilder()
@@ -52,20 +54,14 @@ public class TickBarBuilderTest {
                 .add();
         series.barBuilder()
                 .timePeriod(Duration.ofDays(1))
-                .endTime(now.plus(Duration.ofDays(3)))
-                .closePrice(1)
-                .volume(1)
-                .add();
-        series.barBuilder()
-                .timePeriod(Duration.ofDays(1))
                 .endTime(now.plus(Duration.ofDays(4)))
                 .closePrice(4)
-                .volume(2)
+                .volume(2) // sum is 5, 1 moved to next bar
                 .add();
 
         assertEquals(1, series.getBarCount());
         final var bar = series.getBar(0);
-        assertNumEquals(6, bar.getVolume());
+        assertNumEquals(4, bar.getVolume());
         assertNumEquals(1, bar.getOpenPrice());
         assertNumEquals(4, bar.getClosePrice());
         assertNumEquals(5, bar.getHighPrice());
@@ -87,26 +83,14 @@ public class TickBarBuilderTest {
                 .timePeriod(Duration.ofDays(1))
                 .endTime(now.plus(Duration.ofDays(7)))
                 .closePrice(6)
-                .volume(2)
-                .add();
-        series.barBuilder()
-                .timePeriod(Duration.ofDays(1))
-                .endTime(now.plus(Duration.ofDays(8)))
-                .closePrice(2)
                 .volume(1)
-                .add();
-        series.barBuilder()
-                .timePeriod(Duration.ofDays(1))
-                .endTime(now.plus(Duration.ofDays(9)))
-                .closePrice(5)
-                .volume(2)
                 .add();
         assertEquals(2, series.getBarCount());
 
         final var bar2 = series.getBar(1);
-        assertNumEquals(7, bar2.getVolume());
+        assertNumEquals(4, bar2.getVolume());
         assertNumEquals(2, bar2.getOpenPrice());
-        assertNumEquals(5, bar2.getClosePrice());
+        assertNumEquals(6, bar2.getClosePrice());
         assertNumEquals(6, bar2.getHighPrice());
         assertNumEquals(2, bar2.getLowPrice());
     }
